requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
python-dotenv
Pillow>=10.0
psycopg2-binary
gunicorn